Change long to Long; Add test for domain

+++ url-link/src/test/java/com/leih/url/link/ShortLinkTest.java
@@ -60,7 +60,7 @@ public class ShortLinkTest {
       log.info(code);
       Link shortLink = new Link();
       shortLink.setCode(code);
-      shortLink.setAccountNo(accountNo);
+      shortLink.setAccountNo((long) accountNo);
       shortLink.setSign(CommonUtil.MD5(originalUrl));
       shortLink.setDelete(0);
       shortLinkManager.addShortLink(shortLink);
